+++ src/fhirpath.js
@@ -49,7 +49,7 @@ let datetime  = require("./datetime");
 let logic  = require("./logic");
 const types = require("./types");
 const {
-  FP_DateTime, FP_Time, FP_Quantity,
+  FP_Date, FP_DateTime, FP_Time, FP_Quantity,
   FP_Type, ResourceNode, TypeInfo
 } = types;
 let makeResNode = ResourceNode.makeResNode;
@@ -89,6 +89,7 @@ engine.invocationTable = {
   type:         {fn: types.typeFn, arity: {0: []}},
   ofType:       {fn: filtering.ofTypeFn, arity: {1: ["TypeSpecifier"]}},
   is:           {fn: types.isFn, arity: {1: ["TypeSpecifier"]}},
+  as:           {fn: types.asFn, arity: {1: ["TypeSpecifier"]}},
   tail:         {fn: filtering.tailFn},
   take:         {fn: filtering.takeFn, arity: {1: ["Integer"]}},
   skip:         {fn: filtering.skipFn, arity: {1: ["Integer"]}},
@@ -100,6 +101,7 @@ engine.invocationTable = {
   toInteger:    {fn: misc.toInteger},
   toDecimal:    {fn: misc.toDecimal},
   toString:     {fn: misc.toString},
+  toDate:       {fn: misc.toDate},
   toDateTime:   {fn: misc.toDateTime},
   toTime:       {fn: misc.toTime},
   toBoolean:    {fn: misc.toBoolean},
@@ -108,6 +110,7 @@ engine.invocationTable = {
   convertsToInteger:    {fn: misc.createConvertsToFn(misc.toInteger, 'number')},
   convertsToDecimal:    {fn: misc.createConvertsToFn(misc.toDecimal, 'number')},
   convertsToString:     {fn: misc.createConvertsToFn(misc.toString, 'string')},
+  convertsToDate:       {fn: misc.createConvertsToFn(misc.toDate, FP_Date)},
   convertsToDateTime:   {fn: misc.createConvertsToFn(misc.toDateTime, FP_DateTime)},
   convertsToTime:       {fn: misc.createConvertsToFn(misc.toTime, FP_Time)},
   convertsToQuantity:   {fn: misc.createConvertsToFn(misc.toQuantity, FP_Quantity)},
@@ -162,6 +165,7 @@ engine.invocationTable = {
   "containsOp": {fn: collections.contains,   arity: {2: ["Any", "Any"]}},
   "inOp":       {fn: collections.in,  arity: {2: ["Any", "Any"]}},
   "isOp":       {fn: types.isFn,  arity: {2: ["Any", "TypeSpecifier"]}},
+  "asOp":       {fn: types.asFn,  arity: {2: ["Any", "TypeSpecifier"]}},
   "&":          {fn: math.amp,     arity:  {2: ["String", "String"]}},
   "+":          {fn: math.plus,    arity:  {2: ["Any", "Any"]}, nullable: true},
   "-":          {fn: math.minus,   arity:  {2: ["Any", "Any"]}, nullable: true},
@@ -345,7 +349,7 @@ engine.MemberInvocation = function(ctx, parentData, node ) {
             toAdd = res.data?.[field];
             _toAdd = res.data?.['_' + field];
             if (toAdd !== undefined || _toAdd !== undefined) {
-              childPath = t;
+              childPath += t;
               break;
             }
           }
@@ -360,6 +364,7 @@ engine.MemberInvocation = function(ctx, parentData, node ) {
             childPath = 'Extension';
           }
         }
+        childPath = model && model.path2Type[childPath] || childPath;
 
         if (util.isSome(toAdd) || util.isSome(_toAdd)) {
           if(Array.isArray(toAdd)) {
@@ -587,7 +592,7 @@ engine.evalTable = { // not every evaluator is listed if they are defined on eng
   InvocationExpression: engine.InvocationExpression,
   AdditiveExpression: engine.OpExpression,
   MultiplicativeExpression: engine.OpExpression,
-  TypeExpression: engine.AliasOpExpression({"is": "isOp"}),
+  TypeExpression: engine.AliasOpExpression({"is": "isOp", "as": "asOp"}),
   MembershipExpression: engine.AliasOpExpression({"contains": "containsOp", "in": "inOp"}),
   NullLiteral: engine.NullLiteral,
   EntireExpression: engine.InvocationTerm,
@@ -772,6 +777,8 @@ function compile(path, model, options) {
     return function (fhirData, context) {
       const inObjPath = fhirData && fhirData.__path__;
       const resource = makeResNode(fhirData, path.base || inObjPath);
+      // Globally set model before applying parsed FHIRPath expression
+      TypeInfo.model = model;
       return applyParsedPath(resource, node, context, model, options);
     };
   } else {
@@ -779,6 +786,8 @@ function compile(path, model, options) {
     return function (fhirData, context) {
       const inObjPath = fhirData && fhirData.__path__;
       const resource = inObjPath ? makeResNode(fhirData, inObjPath) : fhirData;
+      // Globally set model before applying parsed FHIRPath expression
+      TypeInfo.model = model;
       return applyParsedPath(resource, node, context, model, options);
     };
   }